split result.rs into module folder
- result/component.rs: GOComponent enum
- result/mueller.rs: Ampl, Mueller types, MuellerMatrix trait, ApproxEq trait
- result/scatt_result.rs: ScatteringBin trait, ScattResult, ScattResult1D/2D
- result/results.rs: Results struct and core implementation
- result/python.rs: Python bindings for Results
- result/integrate.rs: integrate_theta_weighted_component helper
- result.rs: module root with re-exports (modern Rust module style)

Also fix missing settings import in field.rs tests.
